Move some classes out of af-types.h
Move AttributesChangedListener and MarkAttributeDirty out of af-types.h
